LayoutTests/platform/wpe/imported/w3c/web-platform-tests/xhr/send-authentication-basic-setrequestheader-existing-session-expected.txt


FAIL XMLHttpRequest: send() - "Basic" authenticated request using setRequestHeader() when there is an existing session assert_equals: expected "95925d71-bf0c-44a7-ac2c-87c5a9f2896a\npass" but got "defb115b-5535-421f-bfcb-7f14990887b9\nopen-pass"
Note: this test will only work as expected once per browsing session. Restart browser to re-test.
